Add namespace and policy watchers
In order to handle kubernetes network policies, both namespaces
and network policies must be monitored. As policies are always
namespaces, a distinct policy watcher thread is spawn for each
namespace.

This is achieve using a "policy processor". The namespace event
is sent to the policy processor, which starts the policy watcher
thread. The same policy processor will also handle (not in this
patch) translation of k8s network policies into OVN ACLs.

The patch also includes code for parsing the namespace isolation
annotation, and a command line option to enable policies on the
watcher process. Network policies are indeed still a kubernetes
beta API and users might not want to monitor them by default.